[fix] scenario command refactoring & 상속 안받던 버그 수정
