+++ command_core/shell_commands/exit_command.py
@@ -1,0 +1,6 @@
+import sys
+from command_core.base_command import BaseCommand
+
+class ExitCommand(BaseCommand):
+    def execute(self):
+        sys.exit(0)
