+++ command_core/shell_commands/full_read_command.py
@@ -1,7 +1,7 @@
 from ssd_core.abstract_ssd import AbstractSSD
+from command_core.base_command import BaseCommand
 
-
-class FullReadCommand:
+class FullReadCommand(BaseCommand):
     def __init__(self, ssd: AbstractSSD):
         self._ssd = ssd
         self._loop_count = 100
