--- shell_core/commands/exit_command.py
@@ -1,6 +1,0 @@
-import sys
-
-
-class ExitCommand:
-    def execute(self):
-        sys.exit(0)
