--- shell_core/commands/testscenario2.py
@@ -1,52 +1,0 @@
-from ssd_core.abstract_ssd import AbstractSSD
-
-
-class TestScenario2:
-    RESULT_FAIL = "FAIL"
-    RESULT_PASS = "PASS"
-
-    def __init__(self, ssd: AbstractSSD):
-        self._ssd = ssd
-        self.test_constant = 1
-
-    def execute(self):
-        # test scenario from lecture note page #29
-        LOOP_COUNT = 30
-
-        for i in range(LOOP_COUNT):
-            self._ssd.write(4, self.test_constant)
-            self._ssd.write(0, self.test_constant)
-            self._ssd.write(3, self.test_constant)
-            self._ssd.write(1, self.test_constant)
-            self._ssd.write(2, self.test_constant)
-
-            if not self.read_compare(4, self.test_constant):
-                print(TestScenario2.RESULT_FAIL)
-                return
-
-            if not self.read_compare(0, self.test_constant):
-                print(TestScenario2.RESULT_FAIL)
-                return
-
-            if not self.read_compare(3, self.test_constant):
-                print(TestScenario2.RESULT_FAIL)
-                return
-
-            if not self.read_compare(1, self.test_constant):
-                print(TestScenario2.RESULT_FAIL)
-                return
-
-            if not self.read_compare(2, self.test_constant):
-                print(TestScenario2.RESULT_FAIL)
-                return
-
-        print(TestScenario2.RESULT_PASS)
-        return
-
-    def read_compare(self, address, data):
-        try:
-            result = int(self._ssd.read(address), 16)
-
-            return result == data
-        except:
-            return False
