+++ tests/test_shell_full_read_command.py
@@ -1,8 +1,6 @@
-import pytest
-
-from shell_core.commands.full_read_command import FullReadCommand
+from command_core.shell_commands.full_read_command import FullReadCommand
 from ssd_core.abstract_ssd import AbstractSSD
-#from shell_core.commands.full_read_command import FullReadCommand
+#from shell_core.shell_commands.full_read_command import FullReadCommand
 
 def test_full_read_성공(mocker):
     # arrange
