+++ tests/test_shell_write_command.py
@@ -1,6 +1,6 @@
 import pytest
 from ssd_core.abstract_ssd import AbstractSSD
-from shell_core.commands.write_command import WriteCommand
+from command_core.shell_commands.write_command import WriteCommand
 
 
 def test_write_성공(mocker):
